LspProgress is broken
For some reason this API is not working as expected. The autocommand
`LspProgress` is _not_ recognized as a valid event, registering any
handlers is not possible.

I suspect the error comes from somewhere in `src/nvim/api/autocmd.c` the
PR that introduced the changes for this new API is #23958.