make setup exclusive to rust-analyzer
Previously, this plugin would run a setup function for every file type.
This was no problem provided that the file type was not configured to
use LSP, that is, there was no language server associated with this file
type.

However, with the addition of `clangd` into my workflow (i.e. another
LSP that is *not* rust-analyzer), I discovered that running the setup
function produces unwanted errors due to the override of the progress
handler.

Signed-off-by: Oliver T <geronimooliver00@gmail.com>