fix client re-use logic and payload injection
Notes on the payload injection bug. It can happen that the inner inlay
hints handler does not return a payload, say for example when the
request to 'textDocument/InlayHint' fails in which case trying to render
the inlay hints will produce an error.

TODO: the error reporting capabilities need to be better handled. For
now this patch corrects the error, but it would be interesting to let
the user know why the request failed.